fix: remove dependency on node's `_http_common` module
The checks is `isInvalidHeader` method are now backported from node's source.
Both methods were just a regexp checks, so it was trivial to use the very same
logic as in node.

The implementation changed a bit, `checkInvalidHeaderChar` is now implemented via regexp check,
and the `validateHeader` method was split into `validateHeaderName` and `validateHeaderValue`.

Both regexps are taken from the latest code from master. Also verified the exiting test in node,
but they seem to be aligned with the existing test suite (added only one additional assertion
for `undefined`).

Closes #50