fix: correct target stats when sockets are reused (#576)
This PR attempts to fix incorrect stats due to the reuse of target sockets for HTTP(S) protocols. 

Based on https://github.com/apify/proxy-chain/pull/572

Note: I was forced to upgrade `actions/cache` as `v2` was deprecated and it wouldn't run with it.
I also had to edit eslint config to run with `.` instead of `src` and excluded `tests`, because with `src` the CI was failing (no idea why).